graphql: move errors and helpers to errors.go
